rename web/lib/enums to web/lib/enum-helpers
