+++ apps/web/components/icons/ModIconset.tsx
@@ -1,4 +1,4 @@
-import { ModsEnumHelper } from '@/lib/enums';
+import { ModsEnumHelper } from '@/lib/enum-helpers';
 import { Mods } from '@otr/core/osu';
 import { cn } from '@/lib/utils';
 import Image from 'next/image';
