+++ apps/web/components/leaderboard/LeaderboardFilter.tsx
@@ -14,7 +14,7 @@ import { Filter, X } from 'lucide-react';
 import { Popover, PopoverContent, PopoverTrigger } from '../ui/popover';
 import { Input } from '../ui/input';
 import { ToggleGroup, ToggleGroupItem } from '@/components/ui/toggle-group';
-import { RulesetEnumHelper } from '@/lib/enums';
+import { RulesetEnumHelper } from '@/lib/enum-helpers';
 import {
   createSearchParamsFromSchema,
   defaultLeaderboardFilterValues,
