+++ apps/web/components/matches/MatchAdminView.tsx
@@ -4,7 +4,7 @@ import {
   getEnumFlags,
   MatchWarningFlagsEnumHelper,
   MatchRejectionReasonEnumHelper,
-} from '@/lib/enums';
+} from '@/lib/enum-helpers';
 import { matchEditFormSchema } from '@/lib/schema';
 import { cn } from '@/lib/utils';
 import { zodResolver } from '@hookform/resolvers/zod';
