+++ apps/web/components/search/results/PlayerResultContent.tsx
@@ -2,7 +2,7 @@
 
 import { Globe } from 'lucide-react';
 
-import { RulesetEnumHelper } from '@/lib/enums';
+import { RulesetEnumHelper } from '@/lib/enum-helpers';
 import type { PlayerSearchResult } from '@/lib/orpc/schema/search';
 import { highlightMatch } from '@/lib/utils/search';
 import { TierName } from '@/lib/utils/tierData';
