+++ apps/web/components/tournaments/TournamentAdminView.tsx
@@ -45,7 +45,10 @@ import DeleteButton from '../shared/DeleteButton';
 import AcceptPreVerificationStatusesButton from './AcceptPreVerificationStatusesButton';
 import DeleteTournamentBeatmapsButton from './DeleteTournamentBeatmapsButton';
 import { MultipleSelect, Option } from '@/components/select/multiple-select';
-import { getEnumFlags, TournamentRejectionReasonEnumHelper } from '@/lib/enums';
+import {
+  getEnumFlags,
+  TournamentRejectionReasonEnumHelper,
+} from '@/lib/enum-helpers';
 import { TournamentRejectionReason } from '@otr/core/osu';
 import RefetchMatchDataButton from './RefetchMatchDataButton';
 import RefetchBeatmapDataButton from './RefetchBeatmapDataButton';
